requirements.txt
alembic==1.16.5
anyio==4.11.0
colorama==0.4.6
colorlog==6.9.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
joblib==1.5.2
llvmlite==0.45.1
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numba==0.62.1
numpy==1.26.4
openaq==0.4.0
optuna==4.5.0
packaging==25.0
pandas==2.3.3
pathlib==1.0.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.2
six==1.17.0
skforecast==0.18.0
sniffio==1.3.1
SQLAlchemy==2.0.43
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
python-dotenv==1.0.0
affine==2.4.0
asttokens==2.4.1
attrs==24.2.0
blinker==1.9.0
branca==0.8.1
cairocffi==1.7.1
CairoSVG==2.8.2
Cartopy==0.23.0
certifi==2024.8.30
cffi==1.17.1
cftime==1.6.4
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
comm==0.2.2
contourpy==1.2.1
cssselect2==0.7.0
cycler==0.12.1
decorator==5.1.1
defusedxml==0.7.1
executing==2.1.0
fiona==1.9.6
Flask==3.1.2
folium==0.16.0
fonttools==4.53.1
geopandas==0.12.2
gunicorn==23.0.0
idna==3.7
imageio==2.35.1
ipyleaflet==0.20.0
ipython==8.18.1
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.3
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.5.2
jupyter-leaflet==0.20.0
jupyterlab_widgets==3.0.15
kiwisolver==1.4.5
lazy_loader==0.4
lxml==5.3.0
mapbox-vector-tile==2.0.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
netCDF4==1.7.1.post2
networkx==3.2.1
OWSLib==0.30.0
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.43
protobuf==4.25.4
pure_eval==0.2.3
pyclipper==1.3.0.post5
pycparser==2.22
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
pyshp==2.3.1
rasterio==1.3.10
scikit-image==0.24.0
scipy
shapely==2.0.4
stack-data==0.6.3
tifffile==2024.8.30
tinycss2==1.2.1
traitlets==5.14.3
traittypes==0.2.1
urllib3==2.2.2
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
widgetsnbextension==4.0.14
xarray==2024.3.0
xyzservices==2024.6.0
